chaos-agent/requirements.txt
google-genai
requests
matplotlib
numpy
colorama